Updates to handle setting of the SQLite DB location correctly
This script assumes that the file flask-authlib/conf/dev.config.py
has the followiing value of the database URI
SQLALCHEMY_DATABASE_URI = 'sqlite:////<USER_HOME_DIR>/<GALAHAD_HOME_DIR>/flask-authlib/website/sqlite.db'

This will be updated to the correct directory structure based on the current directory path from
which this script will be run.
This will help to be able to clone the repo and run the flask test server in any directory
with no manual changes.