Integrate transducer functionality into Excalibur
This includes all of the transducer API endpoints, as well as a simple CLI
for convenience of testing.  See README.md for details on how to use the CLI.

This functionality has been moved from `galahad/excalibur/`.

Note the config file `.excalibur`, which won't show up with a simple `ls`.