Create a new pki key and copy it over to virtue
Create a new keypair when a new virtue is created.
The name of the key will be based on the virtue ID stored in LDAP
save the keypair in the ~/galahad-keys dir and then scp over the private
key to the virtue in the specific directory that Merlin will look at to
get this key.