Process Kill Actuator (#169)
* Immediate Process Kill Actuator

The processkiller deb file / python script runs as root and waits.  Merlin sends it names of processes to kill.  This needs to be a separate process because Merlin should not run as root but the process killer must.

This also fixes up the netlink communication so that the second half of the process kill actuator can function (sends message to Virtue LSM in the kernel to block any future attempts to start the process).

* Update assembler deb files

* New kernel deb files - new version of LSM with actuator