Added test for constructor and assembler
This will test that constructor creates a unity image and
then using this image the assembler will be called to assemble
a terminal image into the unity image. The resulting image will then
be used to create a role and a virtue from that role.
Verification of the virtue will be added in future iterations.