Updates to make ValorAPI.valor_create() work correctly
- valor.py is now using ssh_tool which be used as a common library for
ssh client
- removed hardcoded stack name from valor.py and obtained stack name
from excalibur node tags added by cloudformation.
- replaced path of cert to standard location where cert is stored in
excalibur node.
- Updated setup.py to only deploy the Valor Router. The valor compute
nodes will be deployed by the ValorManager code, and is not required at
setup time.
- setup.py no longer supports a separate valor setup. The valor setup is
merged with the excalibur setup.
- Fixed setup.sh for router and calor compute nodes. Updates include
formatting changes, addition of br0 router to network interfaces so that
br0 is not created manually on each reboot, setting of IP table rules to
be persistent, cleanup of rc.local to only include commands which are
nto persistent after a reboot, addition of kernel parms to sysctl.conf
file and addition of appropriate comments.
- a separate rc.local file is no longer required. This is being done in
the setup.sh script.