Updated threading in setup and added route on excalibur
- Added non-persistent route on excalibur node to take affect without a
reboot
- Removed threading of excalibur and valor setup as it was causing
issues with the valor setup. Anyhow this setup takes much less time than
creating the ubuntu_pvm image which is already running in a separate
thread at the time valor setup is called. Thus starting of excalibur
setup and of valor setup is nto needed as it does not save on time
during the entire setup process.