Changed the default MTU for eth0 to 1500 (from 9000)
- The default MTU for the ec2 instances is 9000. This is to support
Jumbo Frames. The issue is that when communicating with the virtue
network (guestnet subnet) the OVS devices/interfaces all have a default
MTU of 1500. This causes the SSH connection from excalibur node to the
guestnet subnet (10.91.x.x) network to hang on high network traffic.
Changed the MTU on excalibur to 1500 and this resolves the network
issue.