Updates to test_valor_migrate and associated migrate_virtue method
- Updated test to call the REST API for migrate_virtue. The main reason
for this is that the underlying API call chooses from a available valor
node as the destination valor and can create standby valors in the event
that the current valor count goes below the standby valors count. In
this case the API call previosuly in place in the testcase caused test
ti wait for the threads that started the standby valor creation to
complete. Now these threads will be created in excalibur.
- Updated the REST endpoint in virtue.py to check if
destination_valor_id is specified as a parm in the REST call. If not
then the valor_migrate_virtue will be called with a value of
destination_valor_id of 'None'.
- Updated get_avaialable_valors to return only those standby valors
which are in a 'RUNNING' state.