Updates and fixes in the following places:
- Updated valor_list: This calls Valor Manager's list_valors method now
which will traverse the valor record from rethinkdb and add a virtue
field to each valor entry. This json record will then be returned from
the API method to the end user.
- Updated get_available_valors: This uses list comprehension istead of
the lambda function as it is supposedly more efficient. Also utilize
list_valors method to make it more efficient and modular.
- Updated get_empty_valors: Use list_valors to reuse code and make it
more cleaner.
- Updated destroy_valor: Check if any virtues exist on the valor
associated with the valor_id provided. If any virtues exist then throw
an exception indicating this.