Updated to remove duplicate calls to rethinkdb and add checks feed
- Added checks for add/remove change feed which occurs after the 'galahad'
table is updated during migration. This is a add, remove change feed
followed by a change. The add and remove cause gaius to and create a vm
where one already exists (after migration) and delete a vm where what
has been moved (due to migration).
- Query rethinkdb in constructor to save values of valor_id and guestnet
address which do not change for a valor.
- Store valor_id in history field, instead of valor ID generated by
rethinkdb.