Updates to fix but when source/target valor are same for migration
- Change the call from valor_migrate_virtue to use get_empty_valor such
that if a valor does not exist one will be created on the fly.
- Get the valor ID of the source (virtue's current valor) and ensure
that the destination valor picked is not the same one.
- Utilize a random index when getting a empty valor from a list of
emopty valors. This ensure that the same valor is not always picked if
it is available.
- Throw an error if the destination valor specified by the user is the
same as the source valor for the virtue during migration.