Fixed bug with OAUTH redirect
- For canvas to work the redirect uri needs to be:

redirect = 'https://{}/virtue/test\n' \
           'http://canvas.com:3000/connect/excalibur/callback'.format(self.ip)
This is what the Redirect URI field must be populated with when the APP
is created. However when a access token is required the Redirect URI
supplied has to be:

'https://{}/virtue/test'.format(self.ip)

If it is set to the previous value the redirect is invalid and the
application will fail to get the access token.
As a solution 2 Redirect URIs are defined and only the one with
"canvas.com" is used when creating an OAUTH app.