Fixed virtue application launch to correctly statt the app docker containers
This is an issue with docker where if the docker daemon exits
uncleanly then a system file is locked and docker start fails
with the error:
    Error response from daemon: id already in use
    Error: failed to start containers:
The current workaround is to issue the docker start command
twice. Tne first time it fails with the above error and the
second time it succeeds.