Commented out printing of inodes created as detected by the virtue lsm patch
The Virtue LSM patch prints out events as it detects them through the kernel.
When the MS office docker image is pulled from the docker registry it seems
that there are massive number of inodes created. With the virtue LSM patch
all of these inodes creation is logged in syslog. This seems to over utilize
some system resource and the entire VM gets frozen under this condition.
Commenting out the inode creation `printk` statement in the virtue LSM patch
is a temporary soltion to this issue. This in essense disables the `inode create`
sensor.