Separated out OVS bridge from the rest of setup
- To be able to get a unique free guestnet IP that OVS bridge setup
requires, its setup has been separated. The rest of setup is done during
cloud_init but the OVS bridge setup is done after cloud_init is complete
as it required a guestnet IP which to stay unique needs to be added to
rethinkdb entry as soon as it is obtained as the check for a free
guestnet IP relies on the valid guestnets being in rethinkdb.