Updated to separate OVS bridge setup and add rethinkdb connection reference
- To be able to get a unique free guestnet IP that OVS bridge setup
requires, its setup has been separated. The rest of setup is done during
cloud_init but the OVS bridge setup is done after cloud_init is complete
as it required a guestnet IP which to stay unique needs to be added to
rethinkdb entry as soon as it is obtained as the check for a free
guestnet IP relies on the valid guestnets being in rethinkdb.
- Replaced rethinkdb connection from:
self.connection = rethinkdb.connect(
                host = self.domain_name,
                port = 28015,
                ssl = {
                    'ca_certs':'/var/private/ssl/rethinkdb_cert.pem',
                }).repl()
with:
self.connection = rethinkdb.connect(
                host = self.domain_name,
                port = 28015,
                ssl = {
                    'ca_certs':'/var/private/ssl/rethinkdb_cert.pem',
                })
To not use the REPL default connection as it is not thread safe.
Passed in the connection object everywhere `.run()` was being called on
rethinkdb queries/updates.