valor: add build scripts to generate a virtual appliance
01-make-valor-filesystem.sh : Produces a new filesystem directory
populated with a fresh Ubuntu installation with the XenBlanket
hypervisor and kernel binaries installed.

(A '02' script is not included, but should perform any necessary
customization of the filesystem directory, such as installation
of Valor software, at this point in the build sequence.)

03-generate-appliance.sh : Produces a .ova virtual appliance file,
containing a bootable VM disk image generated by this script and a
virtual machine metadata file with VM configuration parameters.

valor-template.ovf.in : Virtual Appliance template file which is
processed to generate the final virtual machine metadata.
The resulting .ovf file is then included within the .ova output.

Signed-off-by: Christopher Clark <christopher.clark@starlab.io>