Refactored scripts for building of the Xenblanket/Valor disk Image
- moved shell scripts files into compute dir from build
- moved template file to config dir
- Created 02-install-valor.sh to install valor specific files. This
completes the installation of xenbalnket based valor packages. There are
additional components that can be included in this file also i.e
installation of gaius, syslog_ng and OVS networking setup. For now that
is being done through cloud-init of a live system and is being left as
such until it is determined that including it during the build process
of the image is preferred.