Add retry:2 for unit_test to avoid random failure.
If it fails the 1st time, the 2nd will double check actual failure.
To avoid with time/money to issue another deployment just because of
timing issue on first attempt. If it fails 2nd time, then we can assume
an actual bug. Faster double testing than having to deploy a second time
to get a success.